+++ CHANGELOG.md
@@ -2,4 +2,8 @@
 
 ## 0.0.1
 
-Initial version of Glimmer DSL for GTK supporting Hello, World!
+- Initial version of Glimmer DSL for GTK
+- General widget support
+- General widget property support
+- Have `window` automatically call `::Gtk::main_quit` upon hitting `window` close (destroy) button
+- Hello, World! sample